+++ packages/core/src/providers/query/index.tsx
@@ -1,6 +1,6 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { createContext, FC, useContext } from 'react';
-import { useFeatureFlagSubscription } from '../universal/hooks/useFeatureFlagSubscription';
+import { useFeatureFlagSubscription } from '../../universal/hooks/useFeatureFlagSubscription';
 
 const queryClient = new QueryClient();
